LoadState.kt has been moved to ui module
